20221003最新版

+++ src/main/java/controller/AdminCustomerServlet.java
@@ -54,21 +54,42 @@ public class AdminCustomerServlet extends HttpServlet {
 	 */
 	protected void doPost(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
+		CustomerDao customerDao = DaoFactory.createCustomerDao();
+		if (request.getParameter("nameTelMember")==null) {
+			if (request.getParameter("delete").isBlank()) {
+			} else {
+				// Get パラメータの取得
+				String strId = request.getParameter("id");
+				Integer id = Integer.parseInt(strId);
+				// 削除メソッドの引数用に Member オブジェクトを作成
+				Customer customer = new Customer();
+				customer.setId(id);
+				try {
+					// データの削除
 
-		// Get パラメータの取得
-		String strId = request.getParameter("id");
-		Integer id = Integer.parseInt(strId);
-		// 削除メソッドの引数用に Member オブジェクトを作成
-		Customer customer = new Customer();
-		customer.setId(id);
-		try {
-			// データの削除
-			CustomerDao customerDao = DaoFactory.createCustomerDao();
-			customerDao.delete(customer);
-			response.sendRedirect("customer");
+					customerDao.delete(customer);
+					response.sendRedirect("customer");
 
-		} catch (Exception e) {
-			throw new ServletException(e);
+				} catch (Exception e) {
+					throw new ServletException(e);
+				}
+			}
 		}
+
+		else {
+			try {
+				String nameTelMember = request.getParameter("nameTelMember");
+				List<Customer> customerList = customerDao.findByNammeTelMember(nameTelMember);
+				request.setAttribute("customerList", customerList);
+				request.getRequestDispatcher("/WEB-INF/view/admin/customer.jsp").forward(request, response);
+
+			} catch (Exception e) {
+				// TODO 自動生成された catch ブロック
+				e.printStackTrace();
+			}
+
+		}
+
 	}
+
 }
